+++ supertokens_python/recipe/emailpassword/recipe.py
@@ -22,11 +22,9 @@ from supertokens_python.ingredients.emaildelivery.types import \
     EmailDeliveryConfig
 from supertokens_python.normalised_url_path import NormalisedURLPath
 from supertokens_python.recipe.emailpassword.types import (
-    EmailPasswordIngredients, TypeEmailPasswordEmailDeliveryInput)
-from supertokens_python.recipe.emailverification.interfaces import \
-    TypeEmailVerificationEmailDeliveryInput
+    EmailPasswordIngredients, EmailPasswordEmailTemplateVars)
 from supertokens_python.recipe.emailverification.types import \
-    EmailVerificationIngredients
+    EmailVerificationIngredients, VerificationEmailTemplateVars
 from supertokens_python.recipe_module import APIHandled, RecipeModule
 
 from ...exceptions import SuperTokensError
@@ -59,7 +57,7 @@ from .utils import (InputEmailVerificationConfig, InputOverrideConfig,
 class EmailPasswordRecipe(RecipeModule):
     recipe_id = 'emailpassword'
     __instance = None
-    email_delivery: EmailDeliveryIngredient[TypeEmailPasswordEmailDeliveryInput]
+    email_delivery: EmailDeliveryIngredient[EmailPasswordEmailTemplateVars]
 
     def __init__(self, recipe_id: str, app_info: AppInfo,
                  ingredients: EmailPasswordIngredients,
@@ -68,7 +66,7 @@ class EmailPasswordRecipe(RecipeModule):
                  email_verification_feature: Union[InputEmailVerificationConfig, None] = None,
                  override: Union[InputOverrideConfig, None] = None,
                  email_verification_recipe: Union[EmailVerificationRecipe, None] = None,
-                 email_delivery: Union[EmailDeliveryConfig[TypeEmailPasswordEmailDeliveryInput], None] = None,
+                 email_delivery: Union[EmailDeliveryConfig[EmailPasswordEmailTemplateVars], None] = None,
                  ):
         super().__init__(recipe_id, app_info)
         self.config = validate_and_normalise_user_input(self, app_info, sign_up_feature,
@@ -88,7 +86,7 @@ class EmailPasswordRecipe(RecipeModule):
             self.email_verification_recipe = email_verification_recipe
         else:
             ev_email_delivery_ingredient = cast(
-                EmailDeliveryIngredient[TypeEmailVerificationEmailDeliveryInput],
+                EmailDeliveryIngredient[VerificationEmailTemplateVars],
                 self.email_delivery
             )
 
@@ -161,7 +159,7 @@ class EmailPasswordRecipe(RecipeModule):
              reset_password_using_token_feature: Union[InputResetPasswordUsingTokenFeature, None] = None,
              email_verification_feature: Union[InputEmailVerificationConfig, None] = None,
              override: Union[InputOverrideConfig, None] = None,
-             email_delivery: Union[EmailDeliveryConfig[TypeEmailPasswordEmailDeliveryInput], None] = None):
+             email_delivery: Union[EmailDeliveryConfig[EmailPasswordEmailTemplateVars], None] = None):
         def func(app_info: AppInfo):
             if EmailPasswordRecipe.__instance is None:
                 ingredients = EmailPasswordIngredients(None)