+++ supertokens_python/recipe/emailverification/syncio/__init__.py
@@ -15,8 +15,7 @@
 from typing import Any, Dict, Union
 
 from supertokens_python.async_to_sync_wrapper import sync
-from supertokens_python.recipe.emailverification.interfaces import \
-    TypeEmailVerificationEmailDeliveryInput
+from supertokens_python.recipe.emailverification.types import VerificationEmailTemplateVars
 
 
 def create_email_verification_token(
@@ -50,6 +49,6 @@ def revoke_email_verification_tokens(user_id: str, email: str, user_context: Uni
     return sync(revoke_email_verification_tokens(user_id, email, user_context))
 
 
-def send_email(input_: TypeEmailVerificationEmailDeliveryInput, user_context: Union[None, Dict[str, Any]] = None):
+def send_email(input_: VerificationEmailTemplateVars, user_context: Union[None, Dict[str, Any]] = None):
     from supertokens_python.recipe.emailverification.asyncio import send_email
     return sync(send_email(input_, user_context))