+++ supertokens_python/recipe/emailverification/types.py
@@ -11,14 +11,13 @@
 # WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 # License for the specific language governing permissions and limitations
 # under the License.
+from __future__ import annotations
 
-
-from typing import Union
+from typing import Union, Dict, Any
 
 from supertokens_python.ingredients.emaildelivery import \
     EmailDeliveryIngredient
-from supertokens_python.recipe.emailverification.interfaces import \
-    TypeEmailVerificationEmailDeliveryInput
+from supertokens_python.ingredients.emaildelivery.types import SMTPServiceInterface, EmailDeliveryInterface
 
 
 class User:
@@ -27,6 +26,32 @@ class User:
         self.email = email
 
 
+class VerificationEmailTemplateVarsUser:
+    def __init__(self, user_id: str, email: str):
+        self.id = user_id
+        self.email = email
+
+
+class VerificationEmailTemplateVars:
+    def __init__(
+        self,
+        user: VerificationEmailTemplateVarsUser,
+        email_verify_link: str,
+        user_context: Dict[str, Any],
+    ) -> None:
+        self.user = user
+        self.email_verify_link = email_verify_link
+        self.user_context = user_context
+
+
+# Export:
+EmailTemplateVars = VerificationEmailTemplateVars
+
+SMTPOverrideInput = SMTPServiceInterface[EmailTemplateVars]
+
+EmailDeliveryOverrideInput = EmailDeliveryInterface[EmailTemplateVars]
+
+
 class EmailVerificationIngredients:
-    def __init__(self, email_delivery: Union[EmailDeliveryIngredient[TypeEmailVerificationEmailDeliveryInput], None] = None):
+    def __init__(self, email_delivery: Union[EmailDeliveryIngredient[VerificationEmailTemplateVars], None] = None):
         self.email_delivery = email_delivery