+++ supertokens_python/recipe/passwordless/types.py
@@ -15,7 +15,9 @@ from typing import List, Union
 
 from supertokens_python.ingredients.emaildelivery import \
     EmailDeliveryIngredient
+from supertokens_python.ingredients.emaildelivery.types import SMTPServiceInterface, EmailDeliveryInterface
 from supertokens_python.ingredients.smsdelivery import SMSDeliveryIngredient
+from supertokens_python.ingredients.smsdelivery.types import TwilioServiceInterface, SMSDeliveryInterface
 
 
 class User:
@@ -63,7 +65,7 @@ class CreateAndSendCustomEmailParameters:
         self.url_with_link_code: Union[str, None] = url_with_link_code
 
 
-TypePasswordlessEmailDeliveryInput = CreateAndSendCustomEmailParameters
+PasswordlessLoginEmailTemplateVars = CreateAndSendCustomEmailParameters
 
 
 class CreateAndSendCustomTextMessageParameters:
@@ -81,13 +83,24 @@ class CreateAndSendCustomTextMessageParameters:
         self.url_with_link_code: Union[str, None] = url_with_link_code
 
 
-TypePasswordlessSmsDeliveryInput = CreateAndSendCustomTextMessageParameters
+PasswordlessLoginSMSTemplateVars = CreateAndSendCustomTextMessageParameters
 
 
 class PasswordlessIngredients:
     def __init__(self,
-                 email_delivery: Union[EmailDeliveryIngredient[TypePasswordlessEmailDeliveryInput], None] = None,
-                 sms_delivery: Union[SMSDeliveryIngredient[TypePasswordlessSmsDeliveryInput], None] = None,
+                 email_delivery: Union[EmailDeliveryIngredient[PasswordlessLoginEmailTemplateVars], None] = None,
+                 sms_delivery: Union[SMSDeliveryIngredient[PasswordlessLoginSMSTemplateVars], None] = None,
                  ):
         self.email_delivery = email_delivery
         self.sms_delivery = sms_delivery
+
+
+# Export:
+EmailTemplateVars = PasswordlessLoginEmailTemplateVars
+SMSTemplateVars = PasswordlessLoginSMSTemplateVars
+
+SMTPOverrideInput = SMTPServiceInterface[EmailTemplateVars]
+TwilioOverrideInput = TwilioServiceInterface[SMSTemplateVars]
+
+EmailDeliveryOverrideInput = EmailDeliveryInterface[EmailTemplateVars]
+SMSDeliveryOverrideInput = SMSDeliveryInterface[SMSTemplateVars]